Create key_logger.py
How to use pynput to make a Keylogger?

Prerequisites: Python Programming Language
The package pynput.keyboard contains classes for controlling and monitoring the keyboard. pynput is the library of Python that can be used to capture keyboard inputs there the coolest use of this can lie in making keyloggers. The code for the keylogger is given below.

Modules needed

pynput: To install pynput type the below command in the terminal.  

 pip install pynput